move createMinioSecret func to utils file
