+++ packages/scripts/src/helpers/test-runner/index.ts
@@ -21,7 +21,7 @@ import { Kind } from '../kind.js';
 import {
     getArgs, filterBySuite, reportCoverage,
     logE2E, getEnv, groupBySuite,
-    generateTestCaCerts
+    generateTestCaCerts, createMinioSecret
 } from './utils.js';
 import signale from '../signale.js';
 import { getE2EDir, readPackageInfo, listPackages } from '../packages.js';
@@ -30,11 +30,8 @@ import { PublishOptions, PublishType } from '../publish/interfaces.js';
 import { TestTracker } from './tracker.js';
 import {
     MAX_PROJECTS_PER_BATCH, SKIP_DOCKER_BUILD_IN_E2E, TERASLICE_PORT,
-    BASE_DOCKER_IMAGE, K8S_VERSION, NODE_VERSION,
-    ENCRYPT_MINIO
+    BASE_DOCKER_IMAGE, K8S_VERSION, NODE_VERSION, ENCRYPT_MINIO
 } from '../config.js';
-import path from 'node:path';
-import fs from 'node:fs';
 import { K8s } from '../k8s-env/k8s.js';
 
 const logger = debugLogger('ts-scripts:cmd:test');
@@ -299,7 +296,8 @@ async function runE2ETest(
                 // Created a minio secret if needed before helm sync
                 // but after the namespaces have been made
                 if (ENCRYPT_MINIO) {
-                    await createMinioSecret(options);
+                    const k8s = new K8s(TERASLICE_PORT, options.kindClusterName);
+                    await createMinioSecret(k8s);
                 }
 
                 await launchE2EWithHelmfile();
@@ -382,57 +380,4 @@ function printAndGetEnv(suite: string, options: TestOptions) {
         signale.debug(`Setting ${suite} test suite env to ${envStr}`);
     }
     return env;
-}
-
-/**
- * Generates a secret that contains the minio certs and keys
- * used by the minio helm chart when enabling tls.
- * @param options Test options
- */
-async function createMinioSecret(options: TestOptions): Promise<void> {
-    // Grab all needed cert file paths generated by the generateTestCaCerts() function
-    const e2eDir = getE2EDir() as string;
-    const certsDir = path.join(e2eDir, './test/certs/');
-    const privateKeypath = path.join(certsDir, 'private.key');
-    const publicCertPath = path.join(certsDir, 'public.crt');
-    const rootCaPath = path.join(certsDir, '/CAs/rootCA.pem');
-
-    // ensure all necessary files exist
-    if (fs.existsSync(privateKeypath)
-        && fs.existsSync(publicCertPath)
-        && fs.existsSync(rootCaPath)
-    ) {
-        try {
-            // Read in all certs as strings
-            const privateKey = fs.readFileSync(privateKeypath, 'utf-8');
-            const publicCert = fs.readFileSync(publicCertPath, 'utf-8');
-            const rootCA = fs.readFileSync(rootCaPath, 'utf-8');
-
-            // Create a k8s V1Secret object that is used by the k8s client
-            const minioSecret = {
-                metadata: {
-                    name: 'tls-ssl-minio'
-                },
-                stringData: {
-                    'private.key': privateKey,
-                    'public.crt': publicCert,
-                    'rootCA.pem': rootCA
-                },
-                type: 'Opaque'
-            };
-
-            // create k8s client and deploy secret
-            const k8s = new K8s(TERASLICE_PORT, options.kindClusterName);
-            await k8s.createKubernetesSecret('services-dev1', minioSecret);
-        } catch (err) {
-            throw new TSError(`Unable to create minio secret for certificates.\n ${err}`);
-        }
-    } else {
-        throw new Error(
-            'Unable to find all needed minio certificates. One or more paths don\'t exist:\n'
-            + `- ${privateKeypath}\n`
-            + `- ${publicCertPath}\n`
-            + `- ${rootCaPath}\n`
-        );
-    }
 }
