+++ packages/scripts/src/helpers/test-runner/utils.ts
@@ -3,7 +3,7 @@ import { execa } from 'execa';
 import fse from 'fs-extra';
 import {
     debugLogger, get, flatten,
-    isCI, toString
+    isCI, toString, TSError
 } from '@terascope/utils';
 import {
     ArgsMap, ExecEnv, exec,
@@ -15,6 +15,7 @@ import { getServicesForSuite } from '../misc.js';
 import * as config from '../config.js';
 import signale from '../signale.js';
 import { getE2EDir } from '../packages.js';
+import type { K8s } from '../k8s-env/k8s.js';
 
 const logger = debugLogger('ts-scripts:cmd:test');
 
@@ -354,5 +355,57 @@ export async function generateTestCaCerts(): Promise<void> {
         } catch (err) {
             throw new Error(`Error generating ca-certificates for ${serviceList}: ${err.message}`);
         }
+    }
+}
+
+/**
+ * Generates a secret that contains the minio certs and keys
+ * used by the minio helm chart when enabling tls.
+ * @param k8sClient The Kuberetes k8s client
+ */
+export async function createMinioSecret(k8sClient: K8s): Promise<void> {
+    // Grab all needed cert file paths generated by the generateTestCaCerts() function
+    const e2eDir = getE2EDir() as string;
+    const certsDir = path.join(e2eDir, './test/certs/');
+    const privateKeypath = path.join(certsDir, 'private.key');
+    const publicCertPath = path.join(certsDir, 'public.crt');
+    const rootCaPath = path.join(certsDir, '/CAs/rootCA.pem');
+
+    // ensure all necessary files exist
+    if (fse.existsSync(privateKeypath)
+        && fse.existsSync(publicCertPath)
+        && fse.existsSync(rootCaPath)
+    ) {
+        try {
+            // Read in all certs as strings
+            const privateKey = fse.readFileSync(privateKeypath, 'utf-8');
+            const publicCert = fse.readFileSync(publicCertPath, 'utf-8');
+            const rootCA = fse.readFileSync(rootCaPath, 'utf-8');
+
+            // Create a k8s V1Secret object that is used by the k8s client
+            const minioSecret = {
+                metadata: {
+                    name: 'tls-ssl-minio'
+                },
+                stringData: {
+                    'private.key': privateKey,
+                    'public.crt': publicCert,
+                    'rootCA.pem': rootCA
+                },
+                type: 'Opaque'
+            };
+
+            // create k8s client and deploy secret
+            await k8sClient.createKubernetesSecret('services-dev1', minioSecret);
+        } catch (err) {
+            throw new TSError(`Unable to create minio secret for certificates.\n ${err}`);
+        }
+    } else {
+        throw new Error(
+            'Unable to find all needed minio certificates. One or more paths don\'t exist:\n'
+            + `- ${privateKeypath}\n`
+            + `- ${publicCertPath}\n`
+            + `- ${rootCaPath}\n`
+        );
     }
 }
